Make NodeDataType to be extendable by open/closed principle
This is needed when DataType is not a simple data object,
but a larger and smarter object that should not be copied by value.
Having it as a pointer with virtual operators allows the model to
extend it and control the operations done with it (canConnect, etc),
where TypeConverter is not possible to use.