Rely on GraphicsScene selection model.
GraphicsScene knows already how to select the items
when various mouse events occur.
We do not need to manually implement the behavior of the
CTRL key for selecting/deselecting the items.

One thing let to fix was explicit deselection of all the items
when the current item is started being moved by the mouse.