Fix redefined find_package macro for vcpkg (#328)
The original idea was to be able to include the NodeEditor library to
some super-project along with the included Catch2 (also as a
subprobject), i.e.:

```
add_subdirectory(catch2)
add_subdirectory(NodeEditor)
```

To do this we need to detect inside NodeEditor's CMake scripts whether
the Catch2 target was already included. If not, we included our own
Catch2 subdirectory.

The chosen names of the macro caused recursion when called from within
vcpkg.