Drag all the selected models (#357)
The motion of the selected group has been broken in v3.
The original behavior was to employ the default implementation of
QGraphicsItem::mouseMoveEvent that moves all the selected objects
together.

Now the motion is wrapped into a MoveNodeCommand, we need to shift
the selected group manually, one-by-one, by providing new positions to
the Graph Model