Fixed non-emit of BasicGraphicsScene::nodeMoved.
BasicGraphicsScene::nodeMoved was not being emitted anywhere. Now it is
called each time the mouse is released from a node and the dragging
behaviour is stopped.

In the case that Q_EMIT is to be called directly within
BasicGraphicsScene::onNodePositionUpdated, the signal occurs on every
frame update which may be too often. As such, mouse release was chosen
for the emitter.

Signed-off-by: Finn Rayment <finn@rayment.fr>