Fixed non-emit of NodeDelegateModel::in/outputConnectionCreated/Deleted.
The four virtual functions in NodeDelegateModel were not being called
anywhere in code as the logic was removed when FlowScene.cpp was removed
as part of the v3 upgrade.

This functionality has been restored thanks to two new functions
`DataFlowGraphModel::sendConnectionCreation` and
`DataFlowGraphModel::sendConnectionDeletion` which call the slots on
both the input and output nodes of any given connection.

Functions are:
- inputConnectionCreated
- inputConnectionDeleted
- outputConnectionCreated
- outputConnectionDeleted

Signed-off-by: Finn Rayment <finn@rayment.fr>